requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
requests>=2.28.0
flask>=2.0.0
gunicorn>=20.0.0
